+++ ElvUI_LuckyoneUI/Core/Commands.lua
@@ -33,7 +33,7 @@ local AddOns = {
 	ElvUI = true,
 	ElvUI_Libraries = true,
 	ElvUI_LuckyoneUI = true,
-	ElvUI_Options = true
+	ElvUI_Options = true,
 }
 
 -- LuckyoneUI debug mode
@@ -62,12 +62,26 @@ function L1UI:DebugMode(msg)
 	end
 end
 
+-- LuckyoneUI developer switch
+function L1UI:Dev(msg)
+	if msg == 'on' then
+		E.global.L1UI.dev = true
+		L1UI:Print('Developer mode enabled.')
+		E:StaticPopup_Show('L1UI_RL')
+	elseif msg == 'off' then
+		E.global.L1UI.dev = false
+		L1UI:Print('Developer mode disabled.')
+		E:StaticPopup_Show('L1UI_RL')
+	end
+end
+
 -- Register all commands
 function L1UI:LoadCommands()
-	self:RegisterChatCommand('luckyoneui', 'Toggles')
 	self:RegisterChatCommand('luckydebug', 'DebugMode')
+	self:RegisterChatCommand('luckydev', 'Dev')
+	self:RegisterChatCommand('luckyoneui', 'Toggles')
 	if E.Retail then -- Retail chat commands
-		self:RegisterChatCommand('weekly', 'WeeklyRewards')
 		self:RegisterChatCommand('vault', 'WeeklyRewards')
+		self:RegisterChatCommand('weekly', 'WeeklyRewards')
 	end
 end